test(cli/acquisition): hoist CapturingLogger, EnvVarOverride, UnknownPeerProbeResult to TestServices
Move three test helpers that were tucked at the bottom of
InstallationDiscoveryDiscoverAllTests.cs into the shared
tests/Aspire.Cli.Tests/TestServices/ folder alongside the existing
fakes (FakePeerInstallProbe, FakeIdentityChannelReader,
FakeInstallationDiscovery). Namespaces moved to
Aspire.Cli.Tests.TestServices; the test file already imports that
namespace, so call sites are unchanged.

The EnvVarMutatingTestCollection collection definition stays in place
in the test file — collection definitions are conventionally pinned at
the assembly root and have no benefit from being relocated.

This is a pure move with no behavior or signature changes. Sibling test
classes (AspireVersionCheck, DoctorCommandTests) currently reinvent the
env-var-restore pattern; co-locating these helpers with the rest of the
TestServices fakes makes them discoverable for adoption in follow-up
cleanups.
