+++ tests/Aspire.Cli.Tests/TestServices/UnknownPeerProbeResult.cs
@@ -1,0 +1,8 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using Aspire.Cli.Acquisition;
+
+namespace Aspire.Cli.Tests.TestServices;
+
+internal sealed record UnknownPeerProbeResult : PeerProbeResult;
